Rename component to common style
